RF2- added first scheme for controller with focus on user and plates for a particular restaurant

+++ app/Http/Controllers/Admin/PlateController.php
@@ -7,20 +7,26 @@ use App\Http\Requests\StorePlateRequest;
 use App\Http\Requests\UpdatePlateRequest;
 use App\Models\Plate;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class PlateController extends Controller
 {
     public function index(){
-        $plates = Plate::paginate(10);
-        return view('admin.plates.index', compact('plates'));
+        // verififca che il piatto appartenga all'utente
+        $userRestaurant = Auth::user()->restaurant;
+
+        $plates = $userRestaurant->plates;
+        return view('admin.plates.index', compact('plates','userRestaurant'));
     }
     public function create(){
         $plate = new Plate();
-        // manca inserire la parte delle relazioni
         return view('admin.plates.create', compact('plate'));
     }
     public function store(StorePlateRequest $request){
+        $userRestaurant = Auth::user()->restaurant;
+
         $data = $request->validated();
+        $data['restaurant_id'] = $userRestaurant->id;
         $plate = Plate::create($data);
         return redirect()->route('admin.plates.index');
     }
@@ -31,11 +37,24 @@ class PlateController extends Controller
         return view('admin.plates.edit', compact('plate'));
     }
     public function update( UpdatePlateRequest $request, Plate $plate){
+        $userRestaurant = Auth::user()->restaurant;
+
+        if($plate->restaurant_id !== $userRestaurant->id){
+            abort(403);
+        }
+        // recupera dati
         $data = $request->validated();
+        // aggiorna i dati
         $plate->update($data);
-        return redirect()->route('admin.plates.show', $plate);
+        return redirect()->route('admin.plates.index');
     }
     public function destroy(Plate $plate){
+        $userRestaurant = Auth::user()->restaurant;
+
+        if($plate->restaurant_id !== $userRestaurant->id){
+            abort(403);
+        }
+
         $plate->delete();
         return redirect()->route('admin.plates.index');
     }
